api/requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
flask-cors==5.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-SQLAlchemy==3.1.1
greenlet==3.1.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
PyJWT==2.10.1
python-dotenv==1.1.0
SQLAlchemy==2.0.39
typing_extensions==4.12.2
Werkzeug==3.1.3
python-ulid==3.0.0
flask-cors==5.0.1
Pillow==11.2.0
Werkzeug==3.1.3